// src/types/api/reservation-print.ts
export type ReservationPrintProperties = {
  Auth: string | null
  CardType: string | null
  CardNum: string | null
  PaiedAmount: number | null
  CustomerName: string | null
  PartyNo: number | null
  PaymentType: string | null
  Orgin: string | null
  Promo: string | null
  reservationid: string
  showtm: string | null
  TableNums: string | null
  LocAddr1: string | null
  LocCity: string | null
  LocState: string | null
  LocZip: string | null
  LocSName: string | null
  LocURL: string | null
  ShowDate: string | null
  Headliner: string | null
  CheckedIn: number | null
  RefundAmount: number | null
  ReservationTableNumInfo: any[]
  DueAmount: number | null
  PNREF: string | null
  TicketText: string | null
  Address: string | null
  TicketSection: string | null
  TicketDefaultText: string | null
  BottomTicketText: string | null
}